# [CPP Graph] KV-Update Optimization