--- a/intel_extension_for_transformers/llm/runtime/graph/core/ne.h
+++ b/intel_extension_for_transformers/llm/runtime/graph/core/ne.h
@@ -48,6 +48,12 @@
 
 #define NE_SIZE_CALC -1
 
+#if __AVX512F__
+#define NE_ALIGNMENT 64
+#else
+#define NE_ALIGNMENT 32
+#endif
+
 #ifdef __cplusplus
 extern "C" {
 #endif