@DDEle 2023-09-22
> can you try more models to verify the performance gain?

This improvement will only benefits models with fused MHA enabled (i.e. gptj & llama only by now).

I have added some test result for llama in the PR description. Looks that the result is similar with that of gptj.